pyproject.toml
[project]
name = "pilot-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "seaborn>=0.13.2",
    "torch",
    "torchvision",
    "torchaudio",
    "scikit-learn>=1.8.0",
    "python-dotenv>=1.2.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langsmith>=0.7.9",
    "openpyxl>=3.1.5",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
